fix negentropy initial message encoding
