+++ packages/runtime/src/relay-repair.ts
@@ -12,11 +12,14 @@ import {
   filterNegentropyEventRefs,
   reconcileNegentropyRepairSubjects,
   reconcileReplayRepairSubjects,
-  sortNegentropyEventRefsAsc,
   toOrderedEventCursor,
   validateRelayEvent
 } from '@auftakt/core';
 
+import {
+  decodeNegentropyIdListMessage,
+  encodeNegentropyIdListMessage
+} from './negentropy-message.js';
 import { REPAIR_REQUEST_COALESCING_SCOPE } from './request-planning.js';
 
 type RuntimeFilter = Record<string, unknown>;
@@ -107,113 +110,6 @@ function cacheUnsupportedNegentropyRelay(runtime: object, relayUrl: string): voi
 
 function isNegentropyRelayUnsupported(runtime: object, relayUrl: string): boolean {
   return getUnsupportedNegentropyRelayCache(runtime).has(relayUrl);
-}
-
-function encodeHex(bytes: Uint8Array): string {
-  return [...bytes].map((value) => value.toString(16).padStart(2, '0')).join('');
-}
-
-function decodeHex(hex: string): Uint8Array {
-  if (hex.length % 2 !== 0) {
-    throw new Error('negentropy hex payload must have even length');
-  }
-
-  const bytes = new Uint8Array(hex.length / 2);
-  for (let index = 0; index < hex.length; index += 2) {
-    const value = Number.parseInt(hex.slice(index, index + 2), 16);
-    if (!Number.isFinite(value)) {
-      throw new Error('negentropy hex payload contains invalid byte');
-    }
-    bytes[index / 2] = value;
-  }
-  return bytes;
-}
-
-function encodeVarint(value: number): number[] {
-  if (!Number.isInteger(value) || value < 0) {
-    throw new Error('negentropy varint must be a non-negative integer');
-  }
-
-  const digits = [value & 0x7f];
-  let remaining = value >>> 7;
-  while (remaining > 0) {
-    digits.push(remaining & 0x7f);
-    remaining >>>= 7;
-  }
-
-  return digits.reverse().map((digit, index) => (index < digits.length - 1 ? digit | 0x80 : digit));
-}
-
-function decodeVarint(bytes: Uint8Array, start: number): { value: number; next: number } {
-  let value = 0;
-  let index = start;
-
-  while (index < bytes.length) {
-    const byte = bytes[index] ?? 0;
-    value = (value << 7) | (byte & 0x7f);
-    index += 1;
-    if ((byte & 0x80) === 0) {
-      return { value, next: index };
-    }
-  }
-
-  throw new Error('unterminated negentropy varint');
-}
-
-function encodeNegentropyIdListMessage(events: readonly NegentropyEventRef[]): string {
-  const sorted = sortNegentropyEventRefsAsc(events);
-  const bytes: number[] = [0x61, 0x00, 0x00, 0x02, ...encodeVarint(sorted.length)];
-
-  for (const event of sorted) {
-    if (!/^[0-9a-f]{64}$/i.test(event.id)) {
-      throw new Error(`negentropy requires 32-byte hex ids, received: ${event.id}`);
-    }
-    bytes.push(...decodeHex(event.id));
-  }
-
-  return encodeHex(Uint8Array.from(bytes));
-}
-
-function decodeNegentropyIdListMessage(messageHex: string): string[] {
-  const bytes = decodeHex(messageHex);
-  if ((bytes[0] ?? 0) !== 0x61) {
-    throw new Error('unsupported negentropy protocol version');
-  }
-
-  let index = 1;
-  const ids: string[] = [];
-
-  while (index < bytes.length) {
-    const upperTimestamp = decodeVarint(bytes, index);
-    index = upperTimestamp.next;
-    const prefixLength = decodeVarint(bytes, index);
-    index = prefixLength.next + prefixLength.value;
-
-    const mode = decodeVarint(bytes, index);
-    index = mode.next;
-
-    if (mode.value === 0) {
-      continue;
-    }
-
-    if (mode.value !== 2) {
-      throw new Error(`unsupported negentropy mode: ${mode.value}`);
-    }
-
-    const listLength = decodeVarint(bytes, index);
-    index = listLength.next;
-
-    for (let count = 0; count < listLength.value; count += 1) {
-      const nextIndex = index + 32;
-      if (nextIndex > bytes.length) {
-        throw new Error('truncated negentropy id list');
-      }
-      ids.push(encodeHex(bytes.slice(index, nextIndex)));
-      index = nextIndex;
-    }
-  }
-
-  return ids;
 }
 
 function chunkIds(ids: readonly string[], size = 50): RuntimeFilter[] {
